- FIXED: All Python filters are executed with the same Python   interpreter that executes the asciidoc parent (previously filters   were hardwired to execute the 'python' interpreter). This prevents   Python mixups (see http://groups.google.com/group/asciidoc/browse_thread/thread/14e8fcb289a135b/3af3b4e57b827c78?lnk=gst&q=archlinux#3af3b4e57b827c78
- FIXED: Leaky file handles in a2x and music and latex filters which
  created incompatibility problems for Jython. See
http://stackoverflow.com/questions/1832528/is-close-necessary-when-using-iterator-on-a-python-file-object

- FIXED: Jython AsciiDoc API (asciidocapi.py) StringIO incompatibility.

There is still an outstanding Jython related asciidocapi bug that
occurs when using a cStringIO output file: Jython throws an error when
non-ascii characters are found in included CSS or JavaScript files
(e.g. slidy.css), this only happens if you are executing asciidoc via
the asciidocapi and you are using a StringIO output file (as opposed
to a real file).